+++ CHANGELOG.md
@@ -9,8 +9,9 @@ file. This project adheres to [Semantic Versioning](http://semver.org).
 ### 3.1.1
 * minimal Java 9 support ([#7](https://github.com/aburmeis/decs/issues/7))
 * built with Java 11
-* [Appendables](src/main/java/diergo/csv/Appendables.java) are now supported instead of Writer ([#6](https://github.com/aburmeis/decs/issues/6))
+* [Appendables](src/main/java/diergo/csv/Appendables.java) are now supported instead of Writers ([#6](https://github.com/aburmeis/decs/issues/6))
 * migrated tests to [JUnit 5](https://junit.org/junit5/)
+* [Writers](src/main/java/diergo/csv/Writers.java) is deprecated and will be removed in a future release
 
 ### 3.1.0